chore: organize all packets individually
