+++ src/in/inv_button.rs
@@ -1,0 +1,99 @@
+use crate::message::MessageDecoder;
+use crate::packet::Packet;
+use crate::prot::ClientProt;
+use wasm_bindgen::prelude::wasm_bindgen;
+
+#[wasm_bindgen]
+pub struct InvButton {
+    #[wasm_bindgen(readonly)]
+    pub op: u8,
+    #[wasm_bindgen(readonly)]
+    pub obj: u16,
+    #[wasm_bindgen(readonly)]
+    pub slot: u16,
+    #[wasm_bindgen(readonly)]
+    pub component: u16,
+}
+
+#[wasm_bindgen]
+impl InvButton {
+    #[wasm_bindgen(constructor)]
+    pub fn new(
+        op: u8,
+        obj: u16,
+        slot: u16,
+        component: u16,
+    ) -> InvButton {
+        return InvButton {
+            op,
+            obj,
+            slot,
+            component,
+        }
+    }
+}
+
+impl MessageDecoder<InvButton> for InvButton {
+    fn length() -> i32 {
+        return 6;
+    }
+
+    fn decode(prot: ClientProt, buf: &mut Packet) -> InvButton {
+        let op: u8 = match prot {
+            ClientProt::INV_BUTTON1 => 1,
+            ClientProt::INV_BUTTON2 => 2,
+            ClientProt::INV_BUTTON3 => 3,
+            ClientProt::INV_BUTTON4 => 4,
+            ClientProt::INV_BUTTON5 => 5,
+            _ => 0,
+        };
+        return InvButton::new(
+            op,
+            buf.g2(),
+            buf.g2(),
+            buf.g2(),
+        );
+    }
+}
+
+// ----
+
+#[wasm_bindgen]
+pub struct InvButtonD {
+    #[wasm_bindgen(readonly)]
+    pub component: u16,
+    #[wasm_bindgen(readonly)]
+    pub slot: u16,
+    #[wasm_bindgen(readonly)]
+    pub target: u16,
+}
+
+#[wasm_bindgen]
+impl InvButtonD {
+    #[wasm_bindgen(constructor)]
+    pub fn new(
+        component: u16,
+        slot: u16,
+        target: u16,
+    ) -> InvButtonD {
+        return InvButtonD {
+            component,
+            slot,
+            target,
+        }
+    }
+}
+
+impl MessageDecoder<InvButtonD> for InvButtonD {
+    fn length() -> i32 {
+        return 6;
+    }
+
+    fn decode(_: ClientProt, buf: &mut Packet) -> InvButtonD {
+        return InvButtonD::new(
+            buf.g2(),
+            buf.g2(),
+            buf.g2(),
+        );
+    }
+}
