+++ src/in/mod.rs
@@ -1,0 +1,25 @@
+pub mod client_cheat;
+pub mod close_modal;
+pub mod friend;
+pub mod idle_timer;
+pub mod if_button;
+pub mod if_playerdesign;
+pub mod ignore;
+pub mod inv_button;
+pub mod message_private;
+pub mod message_public;
+pub mod move_click;
+pub mod no_timeout;
+pub mod opheld;
+pub mod oploc;
+pub mod opnpc;
+pub mod opobj;
+pub mod opplayer;
+pub mod rebuild_getmaps;
+pub mod resume_pausebutton;
+pub mod resume_countdialog;
+pub mod tutorial_clickside;
+pub mod chat_setmode;
+pub mod reportabuse;
+pub mod event;
+pub mod anticheat;
