+++ src/lib.rs
@@ -2,15 +2,44 @@
 
 use crate::coord::CoordGrid;
 use crate::grid::ZoneMap;
-use crate::info::{NpcInfo, PlayerInfo};
-use crate::message::{AnticheatCycle1, AnticheatCycle2, AnticheatCycle3, AnticheatCycle4, AnticheatCycle5, AnticheatCycle6, AnticheatOp1, AnticheatOp2, AnticheatOp3, AnticheatOp4, AnticheatOp5, AnticheatOp6, AnticheatOp7, AnticheatOp8, AnticheatOp9, ChatSetMode, ClientCheat, CloseModal, EventCameraPosition, EventTracking, FriendListAdd, FriendListDel, IdleTimer, IfButton, IfOpenMain, IfOpenSide, IfPlayerDesign, IgnoreListAdd, IgnoreListDel, IncomingPacket, InfoMessage, InvButton, InvButtonD, MessageDecoder, MessageEncoder, MessageGame, MessagePrivate, MessagePublic, MoveClick, NoTimeout, OpHeld, OpHeldT, OpHeldU, OpLoc, OpLocT, OpLocU, OpNpc, OpNpcT, OpNpcU, OpObj, OpObjT, OpObjU, OpPlayer, OpPlayerT, OpPlayerU, OutgoingPacket, RebuildGetMaps, ReportAbuse, ResumePCountDialog, ResumePauseButton, TutorialClickSide};
+use crate::message::{IncomingPacket, MessageDecoder, OutgoingPacket};
 use crate::npc::Npc;
+use crate::out::if_openmain::IfOpenMain;
+use crate::out::if_openside::IfOpenSide;
+use crate::out::message_game::MessageGame;
+use crate::out::npc_info::NpcInfo;
 use crate::packet::Packet;
 use crate::player::{Chat, ExactMove, Player};
 use crate::prot::{ClientInternalProt, ClientProt};
+use crate::r#in::anticheat::{AnticheatCycle1, AnticheatCycle2, AnticheatCycle3, AnticheatCycle4, AnticheatCycle5, AnticheatCycle6, AnticheatOp1, AnticheatOp2, AnticheatOp3, AnticheatOp4, AnticheatOp5, AnticheatOp6, AnticheatOp7, AnticheatOp8, AnticheatOp9};
+use crate::r#in::chat_setmode::ChatSetMode;
+use crate::r#in::client_cheat::ClientCheat;
+use crate::r#in::close_modal::CloseModal;
+use crate::r#in::event::{EventCameraPosition, EventTracking};
+use crate::r#in::friend::{FriendListAdd, FriendListDel};
+use crate::r#in::idle_timer::IdleTimer;
+use crate::r#in::if_button::IfButton;
+use crate::r#in::if_playerdesign::IfPlayerDesign;
+use crate::r#in::ignore::{IgnoreListAdd, IgnoreListDel};
+use crate::r#in::inv_button::{InvButton, InvButtonD};
+use crate::r#in::message_private::MessagePrivate;
+use crate::r#in::message_public::MessagePublic;
+use crate::r#in::move_click::MoveClick;
+use crate::r#in::no_timeout::NoTimeout;
+use crate::r#in::opheld::{OpHeld, OpHeldT, OpHeldU};
+use crate::r#in::oploc::{OpLoc, OpLocT, OpLocU};
+use crate::r#in::opnpc::{OpNpc, OpNpcT, OpNpcU};
+use crate::r#in::opobj::{OpObj, OpObjT, OpObjU};
+use crate::r#in::opplayer::{OpPlayer, OpPlayerT, OpPlayerU};
+use crate::r#in::rebuild_getmaps::RebuildGetMaps;
+use crate::r#in::reportabuse::ReportAbuse;
+use crate::r#in::resume_countdialog::ResumePCountDialog;
+use crate::r#in::resume_pausebutton::ResumePauseButton;
+use crate::r#in::tutorial_clickside::TutorialClickSide;
 use crate::renderer::{NpcRenderer, PlayerRenderer};
 use crate::visibility::Visibility;
 use once_cell::sync::Lazy;
+use out::player_info::PlayerInfo;
 use std::collections::HashMap;
 use std::ptr::{addr_of, addr_of_mut};
 use wasm_bindgen::prelude::wasm_bindgen;
@@ -23,12 +52,13 @@ mod coord;
 mod player;
 mod prot;
 mod message;
-mod info;
 mod grid;
 mod npc;
 mod visibility;
 mod priority;
 mod category;
+mod r#in;
+mod out;
 
 static mut PLAYERS: Lazy<Vec<Option<Player>>> = Lazy::new(|| vec![None; 2048]);
 static mut PLAYER_GRID: Lazy<HashMap<u32, Vec<i32>>> = Lazy::new(|| HashMap::with_capacity(2048));
